Add 8-hour option to ELO ranker date range filter
- Added "Last 8 hours" option (0.33 days) to date range dropdown
- Updated fetchRecords to handle fractional days using hours calculation
- Improved cache manager logging for fractional day periods
- Changed parseInt to parseFloat for date range values

+++ src/elo-rating-web/app.ts
@@ -184,7 +184,7 @@ class CommunityNotesComparison {
     };
 
     try {
-      const daysBack = parseInt((document.getElementById('dateRange') as HTMLSelectElement).value);
+      const daysBack = parseFloat((document.getElementById('dateRange') as HTMLSelectElement).value);
       
       // Check cache first (unless force refresh)
       const cachedTweets = forceRefresh ? null : this.cacheManager.getFromCache(daysBack);
